fix :wrench: Presonal Option 테이블을 간소화 #11
