+++ packages/server/src/common/enums.ts
@@ -1,0 +1,12 @@
+export enum Environment {
+  Development = "development",
+  Production = "production",
+  Test = "test",
+  Provision = "provision",
+}
+
+export enum OptionType {
+  RADIO = "radio",
+  CHECK = "check",
+  COUNT = "count",
+}
