--- packages/server/src/product-personal-option/dto/create-product-personal-option.dto.ts
@@ -1,1 +1,0 @@
-export class CreateProductPersonalOptionDto {}
