--- packages/server/src/product-personal-option/dto/update-product-personal-option.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateProductPersonalOptionDto } from './create-product-personal-option.dto';
-
-export class UpdateProductPersonalOptionDto extends PartialType(CreateProductPersonalOptionDto) {}
